Move swagger configuration into it's own module
We were on Spring Fox 2 which wasn't compatible with boot 2.x anyway. This upgrades to SpringFox 3 but also moves the entire thing into its own module so people can take it or leave it without setting a property. This will also get it out of our way of conflicting with internal implementations of spring fox that may be modified for netflix specific needs.
