+++ genie-docs/src/docs/asciidoc/_properties.adoc
@@ -613,12 +613,6 @@ of 2.0
 |5000
 |no
 
-|genie.swagger.enabled
-|Whether to enable http://swagger.io/[Swagger] to be bootstrapped into the Genie service so that the endpoint
-/swagger-ui.html shows API documentation generated by the swagger specification
-|false
-|no
-
 |genie.tasks.agent-cleanup.enabled
 |Whether to enable the task that detects jobs whose agent has gone AWOL, and marks them failed
 |true
