+++ genie-swagger/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  com.netflix.genie.swagger.spring.autoconfigure.SwaggerAutoConfiguration
