+++ genie-web/src/main/resources/META-INF/spring.factories
@@ -24,7 +24,6 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.netflix.genie.web.spring.autoconfigure.tasks.node.NodeAutoConfiguration,\
   com.netflix.genie.web.spring.autoconfigure.CachingAutoConfiguration,\
   com.netflix.genie.web.spring.autoconfigure.RetryAutoConfiguration,\
-  com.netflix.genie.web.spring.autoconfigure.SwaggerAutoConfiguration,\
   com.netflix.genie.web.spring.autoconfigure.ValidationAutoConfiguration,\
   com.netflix.genie.web.spring.autoconfigure.ZookeeperAutoConfiguration
 org.springframework.boot.env.EnvironmentPostProcessor=\
